requirements.txt
# Core Python & Config
python-dotenv
pydantic

# API & UI
fastapi
uvicorn
streamlit

# LLM Framework & Orchestration
langchain
langchain-core
langchain-google-genai
langgraph

# Tools & Vector Stores
tavily-python
# L-02: faiss-gpu replaced with faiss-cpu — the vector store is currently a
#        mock; faiss-gpu requires CUDA drivers and fails on CPU-only CI runners.
faiss-cpu

# Security & Middleware
slowapi

# Streaming
sse-starlette

# Resiliency & Tracing
tenacity
langsmith

# Optional data backends (in-memory fallback used when absent)
# redis          # Uncomment for RedisCache (requires REDIS_URL)
# chromadb       # Uncomment for ChromaVectorStore (requires CHROMA_PERSIST_DIR)
